+++ Game Engine/TileMap.cpp
@@ -409,7 +409,7 @@ TILEMAP::TextureSelector::TextureSelector(
 	float bounds_width, float bounds_height,
 	const sf::Texture* texture_Sheet,
 	sf::Font& hide_button_font,
-	int key_time, int max_key_time
+	float key_time, float max_key_time
 ) : keyTime(key_time), maxKeyTime(max_key_time)
 {
 	/*Load Texture Selector Data File*/
@@ -491,7 +491,7 @@ bool TILEMAP::TextureSelector::getKeyTime()
 void TILEMAP::TextureSelector::updateKeyTime(const float& dt)
 {
 	if (this->keyTime < this->maxKeyTime)
-		this->keyTime += static_cast<int>(2.f * dt * (1.f / dt));
+		this->keyTime += 2.f * dt * (1.f / dt);
 }
 void TILEMAP::TextureSelector::update(const sf::Vector2i& mouse_position_window, const float& dt)
 {
